[FE] Fix: HeaderSideBar z-index 수정
- 100 -> 50으로 줄여 play controller보다 위에 위치하는 문제 수정
