+++ client/components/organisms/HeaderSideBar/HeaderSideBar.styles.tsx
@@ -17,8 +17,7 @@ const Logo = styled.h1`
     cursor: pointer;
     display: inline-flex;
     margin: 0 0 5px 0;
-
-    & > * {
+    z & > * {
         margin: 5px;
     }
 `;
@@ -26,16 +25,16 @@ const Logo = styled.h1`
 const SearchContainer = styled.div<SearchContainerProps>`
     overflow: hidden;
     position: fixed;
-    height: ${(props) => props.searching? '67px' : '0'};
+    height: ${(props) => (props.searching ? '67px' : '0')};
     top: 0;
     right: 0;
     left: 225px;
-    z-index: 200;
+    z-index: 100;
     padding: 0 30px;
-    border-bottom: ${(props) => props.searching? '1px solid rgba(0,0,0,.1)' : '0'};
-    background-color: hsla(0,0%,100%,.98);
-    transition: height .2s;
-`
+    border-bottom: ${(props) => (props.searching ? '1px solid rgba(0,0,0,.1)' : '0')};
+    background-color: hsla(0, 0%, 100%, 0.98);
+    transition: height 0.2s;
+`;
 
 const SectionContainer = styled.div`
     padding: 0 15px;
@@ -43,10 +42,10 @@ const SectionContainer = styled.div`
 
 const ProfileContainer = styled.div`
     cursor: pointer;
-    border: 1px solid hsla(0,0%,84.7%,.13);
+    border: 1px solid hsla(0, 0%, 84.7%, 0.13);
     border-width: 1px 0;
     padding: 11px 0;
-`
+`;
 const Header = styled.header`
     position: fixed;
     top: 0;
@@ -54,7 +53,7 @@ const Header = styled.header`
     bottom: 81px;
     width: 225px;
     background-color: black;
-    z-index: 200;
+    z-index: 50;
 `;
 
 export { LogoContainer, Logo, SearchContainer, SectionContainer, ProfileContainer, Header };
