--- client/components/organisms/HeaderSideBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderSideBar';
